+++ core/src/main/java/org/eqasim/core/analysis/pt/PublicTransportLegItem.java
@@ -6,9 +6,10 @@ import org.matsim.pt.transitSchedule.api.TransitLine;
 import org.matsim.pt.transitSchedule.api.TransitRoute;
 import org.matsim.pt.transitSchedule.api.TransitStopFacility;
 
-public class PublicTransportTripItem {
+public class PublicTransportLegItem {
 	public Id<Person> personId;
 	public int personTripId;
+	public int legIndex;
 
 	public Id<TransitStopFacility> accessStopId;
 	public Id<TransitStopFacility> egressStopId;
@@ -16,10 +17,12 @@ public class PublicTransportTripItem {
 	public Id<TransitLine> transitLineId;
 	public Id<TransitRoute> transitRouteId;
 
-	public PublicTransportTripItem(Id<Person> personId, int personTripId, Id<TransitStopFacility> accessStopId,
-			Id<TransitStopFacility> egressStopId, Id<TransitLine> transitLineId, Id<TransitRoute> transitRouteId) {
+	public PublicTransportLegItem(Id<Person> personId, int personTripId, int legIndex,
+			Id<TransitStopFacility> accessStopId, Id<TransitStopFacility> egressStopId, Id<TransitLine> transitLineId,
+			Id<TransitRoute> transitRouteId) {
 		this.personId = personId;
 		this.personTripId = personTripId;
+		this.legIndex = legIndex;
 
 		this.accessStopId = accessStopId;
 		this.egressStopId = egressStopId;